GETP-99 rename: Hashtag와 TechStack을 global.domain.values로 이동
